requirements.txt
requests==2.31.0
opencv-python==4.9.0.80
deepface==0.0.89
numpy==1.26.4
python-dotenv==1.0.0